templates/tensorflow-gpu-1.12/requirements.txt
algorithmia>=1.0.0,<2.0
six
tensorflow-gpu==1.12.0 # if you need a different version, please change this value - bare in mind that it may take longer to compile.